Use price registry as the only source of price (#696)
* add batch prices to price registry

* Have RateLimiter consume prices from PriceRegistry

* fix tests

* update API and deployers

* Remote getTokenPrices from go code

* remove setPrices from go code

* make token prices based on whole token

* remove decimals from PriceRegistry

* port go test changes from PR 698

* prettier

* regenerate wrappers

* no longer use batching in Aggr rate limiter

* remove SerPrices from integration tests

* remove price setting events from AggregateRateLimiter

* address comments

* address comments

* address comments

+++ contracts/src/v0.8/ccip/libraries/Internal.sol
@@ -9,12 +9,12 @@ library Internal {
   struct PriceUpdates {
     TokenPriceUpdate[] tokenPriceUpdates;
     uint64 destChainId; // ------┐ Destination chain Id
-    uint128 usdPerUnitGas; // ---┘ USD per unit of destination chain gas
+    uint192 usdPerUnitGas; // ---┘ USD per unit of destination chain gas
   }
 
   struct TokenPriceUpdate {
     address sourceToken; // Source token
-    uint128 usdPerToken; // USD per unit of token
+    uint192 usdPerToken; // USD per unit of token
   }
 
   struct PoolUpdate {
